Fix Bugs

+++ app/src/main/java/com/nasahacker/nasaeditor/util/FileUtils.kt
@@ -3,14 +3,12 @@ package com.nasahacker.nasaeditor.util
 import android.Manifest
 import android.app.Activity
 import android.content.Context
-import android.content.Intent
 import android.content.pm.PackageManager
 import android.net.Uri
 import android.os.Build
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import androidx.core.content.FileProvider
-import androidx.core.net.toUri
 import com.nasahacker.nasaeditor.util.Constants.APP_FOLDER_NAME
 import com.nasahacker.nasaeditor.util.Constants.FILE_PROVIDER
 import com.nasahacker.nasaeditor.util.Constants.PROJECTS_FOLDER_NAME
@@ -30,7 +28,6 @@ object FileUtils {
                 context,
                 Manifest.permission.WRITE_EXTERNAL_STORAGE
             ) == PackageManager.PERMISSION_GRANTED
-
         }
     }
 
@@ -45,7 +42,6 @@ object FileUtils {
         )
     }
 
-    // Create a new project directory
     fun createProject(context: Context, projectName: String): File {
         val projectFolder = File(
             context.getExternalFilesDir(null),
@@ -57,7 +53,6 @@ object FileUtils {
         return projectFolder
     }
 
-    // Create a file within a specific project
     @Throws(IOException::class)
     fun createFile(
         context: Context,
@@ -72,17 +67,21 @@ object FileUtils {
         return file
     }
 
-    // Write content to a file
     @Throws(IOException::class)
     fun writeToFile(file: File, content: String) {
+        if (file.isDirectory) {
+            throw IOException("Cannot write to a directory: ${file.absolutePath}")
+        }
         FileOutputStream(file).use { fos ->
             fos.write(content.toByteArray())
         }
     }
 
-    // Read content from a file
     @Throws(IOException::class)
     fun readFromFile(file: File): String {
+        if (file.isDirectory) {
+            throw IOException("Cannot read from a directory: ${file.absolutePath}")
+        }
         val content = StringBuilder()
         FileInputStream(file).use { fis ->
             BufferedReader(InputStreamReader(fis)).use { reader ->
@@ -97,26 +96,28 @@ object FileUtils {
 
     @Throws(IOException::class)
     fun readProjectFile(context: Context, projectName: String, fileName: String): String {
-        val projectFolder = createProject(context, projectName) // Create project folder if needed
+        val projectFolder = createProject(context, projectName)
         val file = File(projectFolder, fileName)
         return if (fileExists(file)) {
-            readFromFile(file) // Use existing readFromFile function
+            readFromFile(file)
         } else {
             "" // Handle case where file doesn't exist
         }
     }
 
-    // Update a file with new content (overwrite existing content)
     @Throws(IOException::class)
     fun updateFile(context: Context, projectName: String, fileName: String, newContent: String) {
-        val projectFolder = createProject(context, projectName) // Create project folder if needed
+        val projectFolder = createProject(context, projectName)
         val file = File(projectFolder, fileName)
         writeToFile(file, newContent)
     }
 
-    // Delete a file
     fun deleteFile(context: Context, projectName: String, fileName: String): Boolean {
         val file = File(createProject(context, projectName), fileName)
+        if (file.isDirectory) {
+            // Optionally, you can throw an exception or handle the case where it's a directory
+            return false
+        }
         return file.exists() && file.delete()
     }
 
@@ -125,7 +126,6 @@ object FileUtils {
             context.getExternalFilesDir(null),
             "$APP_FOLDER_NAME/$PROJECTS_FOLDER_NAME/$projectName"
         )
-
         return if (projectFolder.exists()) {
             deleteRecursively(projectFolder)
         } else {
@@ -135,7 +135,6 @@ object FileUtils {
 
     private fun deleteRecursively(file: File): Boolean {
         if (file.isDirectory) {
-            // Delete all files and subdirectories
             val files = file.listFiles()
             files?.forEach { child ->
                 if (!deleteRecursively(child)) {
@@ -146,7 +145,6 @@ object FileUtils {
         return file.delete()
     }
 
-    // Get a list of all files in a specific project
     fun listProjectFiles(context: Context, projectName: String): Array<File>? {
         val projectFolder = File(
             context.getExternalFilesDir(null),
@@ -155,7 +153,6 @@ object FileUtils {
         return projectFolder.listFiles()
     }
 
-    // Get a list of all projects
     fun listProjects(context: Context): Array<File>? {
         val projectsFolder =
             File(context.getExternalFilesDir(null), "$APP_FOLDER_NAME/$PROJECTS_FOLDER_NAME")
@@ -167,7 +164,6 @@ object FileUtils {
         return FileProvider.getUriForFile(context, "${context.packageName}$FILE_PROVIDER", file)
     }
 
-    // Check if a file exists
     fun fileExists(file: File): Boolean {
         return file.exists()
     }
